seirl/debian-tagreadmes/requirements.txt
aiohttp
peewee
python-debian
tqdm
